Solution to the Stem Inc. Energy Prize
Attached is the solution plot to our challenge!

It turns out that this site's load is very power-oriented. With large, short duration peaks. This means that the battery is what we call 'power-limited' and not 'energy-limited', meaning we rarely use the full energy capacity. As a result, the best possible solution was actually equal to the full nameplate power of the battery -- 150kW reduction! 

Awesome job by all Hackers! We were very impressed by your efforts! 

The winners were: 
- Peetz Coffee
- OptimumEnergy

We hope that you enjoyed this challenge a learned a few things about time series forecasting, machine learning, and most importantly, energy storage!

Thanks for helping us hack the clean energy transition!